Consolidate each metric emission based on kubernetes type into single collector, separate collectors and metrics into separate source file by type. Added new node metrics, job metrics, and pvc metrics.
